backend/requirements.txt
alembic==1.14.0
aniso8601==9.0.1
bcrypt==3.2.2
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.15
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.9.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SocketIO==5.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
limits==3.14.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
passlib==1.7.4
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.0
PyJWT==2.10.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
rich==13.9.4
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
WTForms==3.2.1
gunicorn==21.2.0
eventlet==0.35.2
gevent-websocket==0.10.1
